refactor(cli): split run command module
